backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "electronics-marketplace-backend"
version = "0.1.0"
description = "FastAPI backend for the electronics marketplace assignment."
requires-python = ">=3.12"
dependencies = [
  "bcrypt>=4.2,<5.0",
  "email-validator>=2.2,<3.0",
  "fastapi>=0.115,<1.0",
  "PyJWT>=2.9,<3.0",
  "python-multipart>=0.0.9,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
  "pydantic-settings>=2.2,<3.0",
  "sqlalchemy>=2.0,<3.0",
  "pymysql>=1.1,<2.0",
]

[project.scripts]
marketplace-seed = "app.db.seed:main"

[project.optional-dependencies]
dev = [
  "alembic>=1.13,<2.0",
  "httpx>=0.27,<1.0",
  "mypy>=1.10,<2.0",
  "pytest>=8.0,<9.0",
  "pytest-asyncio>=0.23,<1.0",
  "ruff>=0.4,<1.0",
  "types-PyMySQL>=1.1.0.20240524",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
plugins = []
